refactor: add input component
