+++ src/components/index.ts
@@ -1,0 +1,7 @@
+export { Header } from "./header/Header";
+export { Input } from "./input/Input";
+export { RoutesContent } from "./routes.content/RoutesContent";
+export { AddPercentageToTheNumber } from "./percent.calculator/AddPercentageToTheNumber";
+export { SubtractPercentageFromTheNumber } from "./percent.calculator/SubtractPercentageFromTheNumber";
+export { TheNumberPercentage } from "./percent.calculator/TheNumberPercentage";
+export { TheNumberPercentageOfTheNumber } from "./percent.calculator/TheNumberPercentageOfTheNumber";
