--- src/components/routes.content/index.ts
@@ -1,1 +1,0 @@
-export { RoutesContent } from "./RoutesContent";
