+++ src/pages/percent.calculator/PercentCalculator.tsx
@@ -1,7 +1,5 @@
-import { TheNumberPercentage } from './the.number.percentage/TheNumberPercentage'
-import { TheNumberPercentageOfTheNumber } from './the.number.percentage.of.the.number/TheNumberPercentageOfTheNumber'
-import { AddPercentageToTheNumber } from './add.percentage.to.the.number/AddPercentageToTheNumber'
-import { SubtractPercentageFromTheNumber } from './subtract.percentage.from.the.number/SubtractPercentageFromTheNumber'
+
+import { AddPercentageToTheNumber, SubtractPercentageFromTheNumber, TheNumberPercentage, TheNumberPercentageOfTheNumber } from '../../components'
 import './percent.calculator.scss'
 
 export const PercentCalculator = () => {
